made Like Service. but not certainly to do.

+++ src/main/java/com/example/WithPet/repository/Like/LikeRepository.java
@@ -4,13 +4,15 @@ import com.example.WithPet.domain.Like;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
+
 @Repository
 public interface LikeRepository{
     Like save(Like like);
 
     void delete(Like like);
 
-    int findCountByProdId(Long prodId);
+    Long findCountByProdId(Long prodId);
 
     boolean isLike(Long prodId, String userId);
 }
